[change] added script to email report
Modifying view_report() in /maldetect/internals/functions to properly send emails proved problematic when autoupdate_version_hashed="1". The only way to keep any modifications was to disable the integrity check which posses a security risk. While the changes to view_report() implemented functionality beyond cron emails, using custom.cron is a safer alternative if all you want to do is automatically send report emails. 

see: https://serverfault.com/questions/805158/how-to-get-an-email-report-of-whatever-the-most-recent-maldet-scan-is